+++ pkg/controller/backend.go
@@ -19,17 +19,12 @@ package controller
 import (
 	"encoding/json"
 	"fmt"
-	"os"
+	log "github.com/F5Networks/k8s-bigip-ctlr/v3/pkg/vlogger"
+	"k8s.io/apimachinery/pkg/util/intstr"
 	"reflect"
 	"sort"
 	"strconv"
 	"strings"
-	"time"
-
-	rsc "github.com/F5Networks/k8s-bigip-ctlr/v3/pkg/resource"
-	log "github.com/F5Networks/k8s-bigip-ctlr/v3/pkg/vlogger"
-	"github.com/F5Networks/k8s-bigip-ctlr/v3/pkg/writer"
-	"k8s.io/apimachinery/pkg/util/intstr"
 )
 
 const (
@@ -38,9 +33,6 @@ const (
 )
 
 var baseAS3Config = `{
-	"$schema": "https://raw.githubusercontent.com/F5Networks/f5-appsvcs-extension/master/schema/%s/as3-schema-%s.json",
-	"class": "AS3",
-	"declaration": {
 	  "class": "ADC",
 	  "schemaVersion": "%s",
 	  "id": "urn:uuid:85626792-9ee7-46bb-8fc8-4ba708cfdc1d",
@@ -50,637 +42,8 @@ var baseAS3Config = `{
 		 "class": "Controls",
 		 "userAgent": "CIS Configured AS3"
 	  }
-	}
   }
 `
-
-var DEFAULT_PARTITION string
-var DEFAULT_GTM_PARTITION string
-
-func NewAgent(params AgentParams) *Agent {
-	DEFAULT_PARTITION = params.Partition
-	DEFAULT_GTM_PARTITION = params.Partition + "_gtm"
-	postMgr := NewPostManager(params, false)
-	configWriter, err := writer.NewConfigWriter()
-	if nil != err {
-		log.Fatalf("Failed creating ConfigWriter tool: %v", err)
-	}
-	agent := &Agent{
-		PostManager:  postMgr,
-		Partition:    params.Partition,
-		ConfigWriter: configWriter,
-		EventChan:    make(chan interface{}),
-		respChan:     make(chan resourceStatusMeta, 1),
-		userAgent:    params.UserAgent,
-		HttpAddress:  params.HttpAddress,
-		ccclGTMAgent: params.CCCLGTMAgent,
-		disableARP:   params.DisableARP,
-	}
-	// if agent is running in as3 gtm mode and bipip gtm is different from bigip ltm create the gtm agent
-	if isGTMOnSeparateServer(params) {
-		agent.GTMPostManager = NewGTMPostManager(params)
-		go agent.gtmWorker()
-		// retryGTMWorker runs as a separate go routine
-		// blocks on retryChan ; retries failed declarations and polls for accepted tenant statuses
-		go agent.retryGTMWorker()
-	}
-	// agentWorker runs as a separate go routine
-	// blocks on postChan to get new/updated configuration to be posted to BIG-IP
-	go agent.agentWorker()
-
-	// retryWorker runs as a separate go routine
-	// blocks on retryChan ; retries failed declarations and polls for accepted tenant statuses
-	go agent.retryWorker()
-
-	// If running in VXLAN mode, extract the partition name from the tunnel
-	// to be used in configuring a net instance of CCCL for that partition
-	var vxlanPartition string
-	if len(params.VXLANName) > 0 {
-		cleanPath := strings.TrimLeft(params.VXLANName, "/")
-		slashPos := strings.Index(cleanPath, "/")
-		if slashPos == -1 {
-			// No partition
-			vxlanPartition = "Common"
-		} else {
-			// Partition and name
-			vxlanPartition = cleanPath[:slashPos]
-		}
-	}
-	if params.StaticRoutingMode == true {
-		vxlanPartition = params.Partition
-		if params.SharedStaticRoutes == true {
-			vxlanPartition = "Common"
-		}
-	}
-	gs := globalSection{
-		LogLevel:          params.LogLevel,
-		VerifyInterval:    params.VerifyInterval,
-		VXLANPartition:    vxlanPartition,
-		DisableLTM:        true,
-		GTM:               params.CCCLGTMAgent,
-		DisableARP:        params.DisableARP,
-		StaticRoutingMode: params.StaticRoutingMode,
-		MultiClusterMode:  params.MultiClusterMode,
-	}
-
-	// If AS3DEBUG is set, set log level to DEBUG
-	if gs.LogLevel == "AS3DEBUG" {
-		gs.LogLevel = "DEBUG"
-	}
-
-	bs := bigIPSection{
-		BigIPUsername:   params.PostParams.BIGIPUsername,
-		BigIPPassword:   params.PostParams.BIGIPPassword,
-		BigIPURL:        params.PostParams.BIGIPURL,
-		BigIPPartitions: []string{params.Partition},
-	}
-
-	var gtm gtmBigIPSection
-	if len(params.GTMParams.BIGIPURL) == 0 || len(params.GTMParams.BIGIPUsername) == 0 || len(params.GTMParams.BIGIPPassword) == 0 {
-		// gs.GTM = false
-		gtm = gtmBigIPSection{
-			GtmBigIPUsername: params.PostParams.BIGIPUsername,
-			GtmBigIPPassword: params.PostParams.BIGIPPassword,
-			GtmBigIPURL:      params.PostParams.BIGIPURL,
-		}
-		log.Warning("Creating GTM with default bigip credentials as GTM BIGIP Url or GTM BIGIP Username or GTM BIGIP Password is missing on CIS args.")
-	} else {
-		gtm = gtmBigIPSection{
-			GtmBigIPUsername: params.GTMParams.BIGIPUsername,
-			GtmBigIPPassword: params.GTMParams.BIGIPPassword,
-			GtmBigIPURL:      params.GTMParams.BIGIPURL,
-		}
-	}
-	//For IPV6 net config is not required. f5-sdk doesnt support ipv6
-	if !(params.EnableIPV6) {
-		agent.startPythonDriver(
-			gs,
-			bs,
-			gtm,
-			params.PythonBaseDir,
-		)
-	} else {
-		// we only enable metrics as pythondriver is not initialized for ipv6
-		go agent.enableMetrics()
-	}
-	// Set the AS3 version for the LTM Postmanager
-	err = agent.IsBigIPAppServicesAvailable()
-	if err != nil {
-		log.Errorf("%v", err)
-		agent.Stop()
-		os.Exit(1)
-	}
-	// Set the AS3 version on the GTM Postmanager
-	if agent.GTMPostManager != nil {
-		err = agent.GTMPostManager.IsBigIPAppServicesAvailable()
-		if err != nil {
-			log.Errorf("%v", err)
-			agent.Stop()
-			os.Exit(1)
-		}
-	}
-	return agent
-}
-
-func (agent *Agent) Stop() {
-	agent.ConfigWriter.Stop()
-	if !(agent.EnableIPV6) {
-		agent.stopPythonDriver()
-	}
-}
-
-func (agent *Agent) PostConfig(rsConfig ResourceConfigRequest) {
-	// Always push latest activeConfig to channel
-	// Case1: Put latest config into the channel
-	// Case2: If channel is blocked because of earlier config, pop out earlier config and push latest config
-	// Either Case1 or Case2 executes, which ensures the above
-
-	select {
-	case agent.postChan <- rsConfig:
-	case <-agent.postChan:
-		agent.postChan <- rsConfig
-
-	}
-}
-
-// removeDeletedTenantsForBigIP will check the tenant exists on bigip or not
-// if tenant exists and rsConfig does not have tenant, update the tenant with empty PartitionConfig
-func (agent *Agent) removeDeletedTenantsForBigIP(rsConfig *ResourceConfigRequest, cisLabel string) {
-	//Fetching the latest BIGIP Configuration and identify if any tenant needs to be deleted
-	as3Config, err := agent.PostManager.GetAS3DeclarationFromBigIP()
-	if err != nil {
-		log.Errorf("[AS3] Could not fetch the latest AS3 declaration from BIG-IP")
-	}
-	for k, v := range as3Config {
-		if decl, ok := v.(map[string]interface{}); ok {
-			if label, found := decl["label"]; found && label == cisLabel && k != agent.Partition+"_gtm" {
-				if _, ok := rsConfig.ltmConfig[k]; !ok {
-					// adding an empty tenant to delete the tenant from BIGIP
-					priority := 1
-					rsConfig.ltmConfig[k] = &PartitionConfig{Priority: &priority}
-				}
-			}
-		}
-	}
-}
-
-// agentWorker blocks on postChan
-// whenever it gets unblocked, it creates an as3 declaration for modified tenants and posts the request
-func (agent *Agent) agentWorker() {
-	for rsConfig := range agent.postChan {
-		// For the very first post after starting controller, need not wait to post
-		if !agent.firstPost && agent.AS3PostDelay != 0 {
-			// Time (in seconds) that CIS waits to post the AS3 declaration to BIG-IP.
-			log.Debugf("[AS3] Delaying post to BIG-IP for %v seconds ", agent.AS3PostDelay)
-			_ = <-time.After(time.Duration(agent.AS3PostDelay) * time.Second)
-		}
-
-		// If there are no retries going on in parallel, acquiring lock will be straight forward.
-		// Otherwise, we will wait for other workers to complete its current iteration
-		agent.declUpdate.Lock()
-
-		// Fetch the latest config from channel
-		select {
-		case rsConfig = <-agent.postChan:
-		case <-time.After(1 * time.Microsecond):
-		}
-
-		if !(agent.EnableIPV6) && agent.ccclGTMAgent {
-			agent.PostGTMConfig(rsConfig)
-		}
-		// put GMT config in the post channel if gtm agent is running
-		if agent.GTMPostManager != nil {
-			agent.GTMPostManager.PostGTMConfig(rsConfig)
-		}
-
-		decl := agent.createTenantAS3Declaration(rsConfig)
-
-		if len(agent.incomingTenantDeclMap) == 0 {
-			agent.declUpdate.Unlock()
-			continue
-		}
-
-		if agent.HAMode {
-			// if endPoint is not empty means, cis is running in secondary mode
-			// check if the primary cis is up and running
-			if agent.PrimaryClusterHealthProbeParams.EndPointType != "" {
-				if agent.PrimaryClusterHealthProbeParams.statusRunning {
-					// dont post the declaration
-					agent.declUpdate.Unlock()
-					continue
-				} else {
-					if agent.PrimaryClusterHealthProbeParams.statusChanged {
-						agent.PrimaryClusterHealthProbeParams.paramLock.Lock()
-						agent.PrimaryClusterHealthProbeParams.statusChanged = false
-						agent.PrimaryClusterHealthProbeParams.paramLock.Unlock()
-					}
-				}
-			}
-		}
-
-		var updatedTenants []string
-		// initializing the priority tenants
-		var priorityTenants []string
-		/*
-			For every incoming post request, create a new tenantResponseMap.
-			tenantResponseMap will be updated with responses during postConfig.
-			It holds the updatedTenants in the current iteration's as keys.
-			This is needed to update response code in cases (202/404) when httpResponse body does not contain the tenant details.
-		*/
-		agent.tenantResponseMap = make(map[string]tenantResponse)
-
-		for tenant := range agent.incomingTenantDeclMap {
-			// CIS with AS3 doesnt allow write to Common partition.So objects in common partition
-			// should not be updated or deleted by CIS. So removing from tenant map
-			if tenant != "Common" {
-				if _, ok := agent.tenantPriorityMap[tenant]; ok {
-					priorityTenants = append(priorityTenants, tenant)
-				} else {
-					updatedTenants = append(updatedTenants, tenant)
-				}
-				agent.tenantResponseMap[tenant] = tenantResponse{}
-			}
-		}
-
-		// Update the priority tenants first
-		if len(priorityTenants) > 0 {
-			agent.postTenantsDeclaration(decl, rsConfig, priorityTenants)
-		}
-		// Updating the remaining tenants
-		agent.postTenantsDeclaration(decl, rsConfig, updatedTenants)
-
-		agent.declUpdate.Unlock()
-	}
-}
-
-// Post the tenants declaration
-func (agent *Agent) postTenantsDeclaration(decl as3Declaration, rsConfig ResourceConfigRequest, tenants []string) {
-	cfg := agentConfig{
-		data:      string(decl),
-		as3APIURL: agent.getAS3APIURL(tenants),
-		id:        rsConfig.reqId,
-	}
-
-	agent.publishConfig(cfg)
-
-	// Don't update ARPs if disableARP is set to true
-	if !agent.disableARP {
-		go agent.updateARPsForPoolMembers(rsConfig)
-	}
-
-	agent.updateTenantResponseMap(true)
-
-	if len(agent.retryTenantDeclMap) > 0 {
-		// Activate retry
-		select {
-		case agent.retryChan <- struct{}{}:
-		case <-agent.retryChan:
-			agent.retryChan <- struct{}{}
-		}
-	}
-
-	/*
-		If there are any tenants with 201 response code,
-		poll for its status continuously and block incoming requests
-	*/
-	agent.pollTenantStatus()
-
-	// notify resourceStatusUpdate response handler on successful tenant update
-	agent.notifyRscStatusHandler(cfg.id, true)
-}
-
-func (agent *Agent) notifyRscStatusHandler(id int, overwriteCfg bool) {
-
-	rscUpdateMeta := resourceStatusMeta{
-		id,
-		make(map[string]struct{}),
-	}
-	for tenant := range agent.retryTenantDeclMap {
-		rscUpdateMeta.failedTenants[tenant] = struct{}{}
-	}
-	// If triggerred from retry block, process the previous successful request completely
-	if !overwriteCfg {
-		agent.respChan <- rscUpdateMeta
-	} else {
-		// Always push latest id to channel
-		// Case1: Put latest id into the channel
-		// Case2: If channel is blocked because of earlier id, pop out earlier id and push latest id
-		// Either Case1 or Case2 executes, which ensures the above
-		select {
-		case agent.respChan <- rscUpdateMeta:
-		case <-agent.respChan:
-			agent.respChan <- rscUpdateMeta
-		}
-	}
-}
-
-func (agent *Agent) updateARPsForPoolMembers(rsConfig ResourceConfigRequest) {
-	allPoolMembers := rsConfig.ltmConfig.GetAllPoolMembers()
-
-	// Convert allPoolMembers to rsc.Members so that vxlan Manger accepts
-	var allPoolMems []rsc.Member
-
-	for _, poolMem := range allPoolMembers {
-		allPoolMems = append(
-			allPoolMems,
-			rsc.Member(poolMem),
-		)
-	}
-	if agent.EventChan != nil {
-		select {
-		case agent.EventChan <- allPoolMems:
-			log.Debugf("Controller wrote endpoints to VxlanMgr")
-		case <-time.After(3 * time.Second):
-		}
-	}
-}
-
-// retryWorker blocks on retryChan
-// whenever it gets unblocked, retries failed declarations and polls for accepted tenant statuses
-func (agent *Agent) retryWorker() {
-
-	/*
-		retryWorker runs as a goroutine. It is idle until an arrives at retryChan.
-		retryTenantDeclMal holds all information about tenant adc configuration and response codes.
-
-		Once retryChan is signalled, retryWorker posts tenant declarations and/or polls for accepted tenants' statuses continuously until it succeeds
-		Locks are used to block retries if an incoming request arrives at agentWorker.
-
-		For each iteration, retryWorker tries to acquire agent.declUpdate lock.
-		During an ongoing agentWorker's activity, retryWorker tries to wait until agent.declUpdate lock is acquired
-		Similarly, during an ongoing retry, agentWorker waits for graceful termination of ongoing iteration - i.e., until agent.declUpdate is unlocked
-
-	*/
-
-	for range agent.retryChan {
-
-		for len(agent.retryTenantDeclMap) != 0 {
-
-			if agent.HAMode {
-				// if endPoint is not empty -> cis is running in secondary mode
-				// check if the primary cis is up and running
-				if agent.PrimaryClusterHealthProbeParams.EndPointType != "" {
-					if agent.PrimaryClusterHealthProbeParams.statusRunning {
-						agent.retryTenantDeclMap = make(map[string]*tenantParams)
-						// dont post the declaration
-						continue
-					}
-				}
-			}
-
-			agent.declUpdate.Lock()
-
-			// If we had a delay in acquiring lock, re-check if we have any tenants to be retried
-			if len(agent.retryTenantDeclMap) == 0 {
-				agent.declUpdate.Unlock()
-				break
-			}
-
-			log.Debugf("[AS3] Posting failed tenants configuration in %v seconds", timeoutMedium)
-
-			//If there are any 201 tenants, poll for its status
-			agent.pollTenantStatus()
-
-			//If there are any failed tenants, retry posting them
-			agent.retryFailedTenant(agent.userAgent)
-
-			agent.notifyRscStatusHandler(0, false)
-
-			agent.declUpdate.Unlock()
-		}
-	}
-}
-
-func (agent *Agent) PostGTMConfig(config ResourceConfigRequest) {
-
-	dnsConfig := make(map[string]interface{})
-	wideIPs := WideIPs{}
-
-	for _, gtmPartitionConfig := range config.gtmConfig {
-		for _, v := range gtmPartitionConfig.WideIPs {
-			wideIPs.WideIPs = append(wideIPs.WideIPs, v)
-		}
-	}
-
-	dnsConfig["Common"] = wideIPs
-	doneCh, errCh, err := agent.ConfigWriter.SendSection("gtm", dnsConfig)
-
-	if nil != err {
-		log.Warningf("Failed to write gtm config section: %v", err)
-	} else {
-		select {
-		case <-doneCh:
-			log.Debugf("Wrote gtm config section: %v", config.gtmConfig)
-		case e := <-errCh:
-			log.Warningf("Failed to write gtm config section: %v", e)
-		case <-time.After(time.Second):
-			log.Warningf("Did not receive write response in 1s")
-		}
-	}
-}
-
-// Creates AS3 adc only for tenants with updated configuration
-func (agent *Agent) createTenantAS3Declaration(config ResourceConfigRequest) as3Declaration {
-	// Re-initialise incomingTenantDeclMap map and tenantPriorityMap for each new config request
-	agent.incomingTenantDeclMap = make(map[string]as3Tenant)
-	agent.tenantPriorityMap = make(map[string]int)
-	for tenant, cfg := range agent.createAS3LTMAndGTMConfigADC(config) {
-		if !reflect.DeepEqual(cfg, agent.cachedTenantDeclMap[tenant]) ||
-			(agent.PrimaryClusterHealthProbeParams.EndPoint != "" && agent.PrimaryClusterHealthProbeParams.statusChanged) {
-			agent.incomingTenantDeclMap[tenant] = cfg.(as3Tenant)
-		} else {
-			// cachedTenantDeclMap always holds the current configuration on BigIP(lets say A)
-			// When an invalid configuration(B) is reverted (to initial A) (i.e., config state A -> B -> A),
-			// delete entry from retryTenantDeclMap if any
-			delete(agent.retryTenantDeclMap, tenant)
-			// Log only when it's primary/standalone CIS or when it's secondary CIS and primary CIS is down
-			if agent.PrimaryClusterHealthProbeParams.EndPoint == "" || !agent.PrimaryClusterHealthProbeParams.statusRunning {
-				log.Debugf("[AS3] No change in %v tenant configuration", tenant)
-			}
-		}
-	}
-
-	// gtmAS3
-	//gtmPartitionConfig := agent.createAS3GTMConfigADC(config)
-	//if gtmPartitionConfig != nil {
-	//	if !reflect.DeepEqual(gtmPartitionConfig, agent.cachedTenantDeclMap[gtmPartition]) {
-	//		agent.incomingTenantDeclMap[gtmPartition] = gtmPartitionConfig
-	//	} else {
-	//		delete(agent.retryTenantDeclMap, gtmPartition)
-	//	}
-	//}
-
-	return agent.createAS3Declaration(agent.incomingTenantDeclMap, agent.userAgent)
-}
-
-func (agent *Agent) createAS3LTMAndGTMConfigADC(config ResourceConfigRequest) as3ADC {
-	adc := agent.createAS3LTMConfigADC(config)
-	if !agent.ccclGTMAgent && agent.GTMPostManager == nil {
-		adc = agent.createAS3GTMConfigADC(config, adc)
-	}
-	return adc
-}
-
-func (agent *Agent) createAS3GTMConfigADC(config ResourceConfigRequest, adc as3ADC) as3ADC {
-	if len(config.gtmConfig) == 0 {
-		sharedApp := as3Application{}
-		sharedApp["class"] = "Application"
-		sharedApp["template"] = "shared"
-		cisLabel := agent.Partition
-		tenantDecl := as3Tenant{
-			"class":              "Tenant",
-			as3SharedApplication: sharedApp,
-			"label":              cisLabel,
-		}
-		adc[DEFAULT_GTM_PARTITION] = tenantDecl
-
-		return adc
-	}
-
-	for pn, gtmPartitionConfig := range config.gtmConfig {
-		var tenantDecl as3Tenant
-		var sharedApp as3Application
-
-		if obj, ok := adc[pn]; ok {
-			tenantDecl = obj.(as3Tenant)
-			sharedApp = tenantDecl[as3SharedApplication].(as3Application)
-		} else {
-			sharedApp = as3Application{}
-			sharedApp["class"] = "Application"
-			sharedApp["template"] = "shared"
-
-			tenantDecl = as3Tenant{
-				"class":              "Tenant",
-				as3SharedApplication: sharedApp,
-			}
-		}
-
-		for domainName, wideIP := range gtmPartitionConfig.WideIPs {
-
-			gslbDomain := as3GLSBDomain{
-				Class:              "GSLB_Domain",
-				DomainName:         wideIP.DomainName,
-				RecordType:         wideIP.RecordType,
-				LBMode:             wideIP.LBMethod,
-				PersistenceEnabled: wideIP.PersistenceEnabled,
-				PersistCidrIPv4:    wideIP.PersistCidrIPv4,
-				PersistCidrIPv6:    wideIP.PersistCidrIPv6,
-				TTLPersistence:     wideIP.TTLPersistence,
-				Pools:              make([]as3GSLBDomainPool, 0, len(wideIP.Pools)),
-			}
-			if wideIP.ClientSubnetPreferred != nil {
-				gslbDomain.ClientSubnetPreferred = wideIP.ClientSubnetPreferred
-			}
-			for _, pool := range wideIP.Pools {
-				gslbPool := as3GSLBPool{
-					Class:          "GSLB_Pool",
-					RecordType:     pool.RecordType,
-					LBMode:         pool.LBMethod,
-					LBModeFallback: pool.LBModeFallBack,
-					Members:        make([]as3GSLBPoolMemberA, 0, len(pool.Members)),
-					Monitors:       make([]as3ResourcePointer, 0, len(pool.Monitors)),
-				}
-
-				for _, mem := range pool.Members {
-					gslbPool.Members = append(gslbPool.Members, as3GSLBPoolMemberA{
-						Enabled: true,
-						Server: as3ResourcePointer{
-							BigIP: pool.DataServer,
-						},
-						VirtualServer: mem,
-					})
-				}
-
-				for _, mon := range pool.Monitors {
-					gslbMon := as3GSLBMonitor{
-						Class:    "GSLB_Monitor",
-						Interval: mon.Interval,
-						Type:     mon.Type,
-						Send:     mon.Send,
-						Receive:  mon.Recv,
-						Timeout:  mon.Timeout,
-					}
-
-					gslbPool.Monitors = append(gslbPool.Monitors, as3ResourcePointer{
-						Use: mon.Name,
-					})
-
-					sharedApp[mon.Name] = gslbMon
-				}
-				gslbDomain.Pools = append(gslbDomain.Pools, as3GSLBDomainPool{Use: pool.Name, Ratio: pool.Ratio})
-				sharedApp[pool.Name] = gslbPool
-			}
-
-			sharedApp[strings.Replace(domainName, "*", "wildcard", -1)] = gslbDomain
-		}
-		adc[pn] = tenantDecl
-	}
-
-	return adc
-}
-
-func (agent *Agent) createAS3LTMConfigADC(config ResourceConfigRequest) as3ADC {
-	adc := as3ADC{}
-	cisLabel := agent.Partition
-
-	if agent.HAMode {
-		// Delete the tenant which is monitored by CIS and current request does not contain it, if it's the first post or
-		// if it's secondary CIS and primary CIS is down and statusChanged is true
-		if agent.firstPost ||
-			(agent.PrimaryClusterHealthProbeParams.EndPoint != "" && !agent.PrimaryClusterHealthProbeParams.statusRunning &&
-				agent.PrimaryClusterHealthProbeParams.statusChanged) {
-			agent.removeDeletedTenantsForBigIP(&config, cisLabel)
-			agent.firstPost = false
-		}
-	}
-
-	for tenant := range agent.cachedTenantDeclMap {
-		if _, ok := config.ltmConfig[tenant]; !ok && !agent.isGTMTenant(tenant) {
-			// Remove partition
-			adc[tenant] = getDeletedTenantDeclaration(agent.Partition, tenant, cisLabel)
-		}
-	}
-	for tenantName, partitionConfig := range config.ltmConfig {
-		// TODO partitionConfig priority can be overridden by another request if agent is unable to process the prioritized request in time
-		partitionConfig.PriorityMutex.RLock()
-		if *(partitionConfig.Priority) > 0 {
-			agent.tenantPriorityMap[tenantName] = *(partitionConfig.Priority)
-		}
-		partitionConfig.PriorityMutex.RUnlock()
-		if len(partitionConfig.ResourceMap) == 0 {
-			// Remove partition
-			adc[tenantName] = getDeletedTenantDeclaration(agent.Partition, tenantName, cisLabel)
-			continue
-		}
-		// Create Shared as3Application object
-		sharedApp := as3Application{}
-		sharedApp["class"] = "Application"
-		sharedApp["template"] = "shared"
-
-		// Process rscfg to create AS3 Resources
-		processResourcesForAS3(partitionConfig.ResourceMap, sharedApp, config.shareNodes, tenantName)
-
-		// Process CustomProfiles
-		processCustomProfilesForAS3(partitionConfig.ResourceMap, sharedApp, agent.bigIPAS3Version)
-
-		// Process Profiles
-		processProfilesForAS3(partitionConfig.ResourceMap, sharedApp)
-
-		processIRulesForAS3(partitionConfig.ResourceMap, sharedApp)
-
-		processDataGroupForAS3(partitionConfig.ResourceMap, sharedApp)
-
-		// Create AS3 Tenant
-		tenantDecl := as3Tenant{
-			"class":              "Tenant",
-			"defaultRouteDomain": config.defaultRouteDomain,
-			as3SharedApplication: sharedApp,
-			"label":              cisLabel,
-		}
-		adc[tenantName] = tenantDecl
-	}
-	return adc
-}
 
 func getDeletedTenantDeclaration(defaultPartition, tenant, cisLabel string) as3Tenant {
 	if defaultPartition == tenant {
