+++ pkg/controller/informers_test.go
@@ -374,11 +374,11 @@ var _ = Describe("Informers Tests", func() {
 			}
 			mockCtlr.Agent = &Agent{
 				PostManager: &PostManager{
-					postChan: make(chan ResourceConfigRequest, 1),
 					PostParams: PostParams{
 						BIGIPURL: "10.10.10.1",
 					},
 				},
+				reqChan: make(chan ResourceConfigRequest, 1),
 			}
 
 			mockCtlr.requestQueue = &requestQueue{sync.Mutex{}, list.New()}
